select.d: replace SelectSelector with PollSelector
Due to the way SocketSet is implemented in std.socket, the time
complexity of our SelectSelector is O(n^2) on Windows, which is
unacceptable as the number of connections grows.

While it's not recommended to run Soulfind on Windows if you want
to handle a large number of connections efficiently, we can use
poll() / WSAPoll() instead of select(), which is good enough in
most cases, and slightly better than select() on less common POSIX
systems.