Move user privilege commands to soulsetup
These are system-level commands that have more lasting effects,
so move them to soulsetup instead of keeping them as admin commands.
This change also brings the benefit of being able to inform
connected users when their privileges expire, and the possibility
of granting privileges with external tools by modifying the database.